requirements.txt
alabaster==0.7.13
asgiref==3.7.2
astroid==3.0.1
Babel==2.13.0
beautifulsoup4==4.12.2
bootstrap-py==1.0.0
bootstrap4==0.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
colorama==0.4.6
coverage==7.3.2
crispy-bootstrap5==0.7
cryptography==41.0.4
defusedxml==0.7.1
dill==0.3.7
Django==4.2.6
django-allauth==0.57.0
django-bootstrap-v5==1.0.11
django-crispy-forms==2.1
docutils==0.20.1
facebook-sdk==3.1.0
gitdb==4.0.10
GitPython==3.1.37
gunicorn==21.2.0
idna==3.4
imagesize==1.4.1
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pguard==0.4.0
Pillow==10.0.1
platformdirs==3.11.0
pluggy==1.3.0
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
pylint==3.0.2
pytest==7.4.2
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==68.2.2
setuptools-scm==8.0.4
smmap==5.0.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
soupsieve==2.5
sqlparse==0.4.4
tomlkit==0.12.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
